rename greenline specific stencils files
